pyproject.toml
[project]
name = "stash-plugins"
version = "0.1.0"
description = "Utilities for the stash-plugins repository"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "graphql-core>=3.2.6",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.31.0",
    "ruff>=0.11.8",
]

[project.scripts]
download-stash-schema = "stash_tools.cli:main"
fix = "stash_tools.dev:fix_main"
lint = "stash_tools.dev:lint_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/stash_tools"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "UP",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.basedpyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"
include = [
    "src",
    "plugins",
]
exclude = [
    "**/__pycache__",
    ".venv",
]
reportMissingTypeStubs = false
